# openapi.json
{
 "components": {
  "schemas": {
...
   "Body_upload_document_v1_admin_kb_documents_post": {
    "properties": {
     "file": {
      "contentMediaType": "application/octet-stream",
      "title": "File",
      "type": "string"
     },
     "meta": {
      "title": "Meta",
      "type": "string"
     }
    },
    "required": [
     "file",
     "meta"
    ],
    "title": "Body_upload_document_v1_admin_kb_documents_post",
    "type": "object"
   },
...
   "ClausePatch": {
    "additionalProperties": false,
    "properties": {
     "body": {
      "anyOf": [
       {
        "maxLength": 20000,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Body"
     },
     "clause_no": {
      "anyOf": [
       {
        "maxLength": 40,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Clause No"
     },
     "excluded": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Excluded"
     }
    },
    "title": "ClausePatch",
    "type": "object"
   },
...
   "DocPatch": {
    "additionalProperties": false,
    "properties": {
     "categories": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Categories"
     },
     "meta": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Meta"
     },
     "org_codes": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Org Codes"
     },
     "work_types": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Work Types"
     }
    },
    "title": "DocPatch",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "RefItem": {
    "additionalProperties": false,
    "properties": {
     "clause_no": {
      "maxLength": 40,
      "title": "Clause No",
      "type": "string"
     },
     "doc_code": {
      "maxLength": 60,
      "title": "Doc Code",
      "type": "string"
     },
     "rule_code": {
      "maxLength": 20,
      "title": "Rule Code",
      "type": "string"
     }
    },
    "required": [
     "rule_code",
     "doc_code",
     "clause_no"
    ],
    "title": "RefItem",
    "type": "object"
   },
   "RefsBody": {
    "additionalProperties": false,
    "properties": {
     "refs": {
      "items": {
       "$ref": "#/components/schemas/RefItem"
      },
      "maxItems": 2000,
      "title": "Refs",
      "type": "array"
     }
    },
    "required": [
     "refs"
    ],
    "title": "RefsBody",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "AI check of work permit section 2",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/v1/admin/kb/clauses/{clause_id}": {
   "patch": {
    "operationId": "patch_clause_v1_admin_kb_clauses__clause_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "clause_id",
      "required": true,
      "schema": {
       "title": "Clause Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ClausePatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "string"
         },
         "title": "Response Patch Clause V1 Admin Kb Clauses  Clause Id  Patch",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Clause"
   }
  },
  "/v1/admin/kb/documents": {
   "get": {
    "operationId": "list_documents_v1_admin_kb_documents_get",
    "parameters": [
     {
      "in": "query",
      "name": "org",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Org"
      }
     },
     {
      "in": "query",
      "name": "category",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Category"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "title": "Response List Documents V1 Admin Kb Documents Get"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Documents"
   },
   "post": {
    "operationId": "upload_document_v1_admin_kb_documents_post",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_document_v1_admin_kb_documents_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Upload Document V1 Admin Kb Documents Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Upload Document"
   }
  },
  "/v1/admin/kb/documents/{doc_id}": {
   "get": {
    "operationId": "get_document_v1_admin_kb_documents__doc_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "doc_id",
      "required": true,
      "schema": {
       "title": "Doc Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "title": "Response Get Document V1 Admin Kb Documents  Doc Id  Get"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Document"
   },
   "patch": {
    "operationId": "patch_document_v1_admin_kb_documents__doc_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "doc_id",
      "required": true,
      "schema": {
       "title": "Doc Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DocPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "string"
         },
         "title": "Response Patch Document V1 Admin Kb Documents  Doc Id  Patch",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Document"
   }
  },
  "/v1/admin/kb/documents/{doc_id}/activate": {
   "post": {
    "operationId": "activate_document_v1_admin_kb_documents__doc_id__activate_post",
    "parameters": [
     {
      "in": "path",
      "name": "doc_id",
      "required": true,
      "schema": {
       "title": "Doc Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Activate Document V1 Admin Kb Documents  Doc Id  Activate Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Activate Document"
   }
  },
  "/v1/admin/kb/documents/{doc_id}/diff": {
   "get": {
    "operationId": "document_diff_v1_admin_kb_documents__doc_id__diff_get",
    "parameters": [
     {
      "in": "path",
      "name": "doc_id",
      "required": true,
      "schema": {
       "title": "Doc Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "title": "Response Document Diff V1 Admin Kb Documents  Doc Id  Diff Get"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Document Diff"
   }
  },
  "/v1/admin/kb/rule-refs": {
   "put": {
    "operationId": "put_rule_refs_v1_admin_kb_rule_refs_put",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RefsBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "integer"
         },
         "title": "Response Put Rule Refs V1 Admin Kb Rule Refs Put",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Rule Refs"
   }
  },
  "/v1/admin/kb/rule-refs/broken": {
   "get": {
    "operationId": "broken_refs_v1_admin_kb_rule_refs_broken_get",
    "parameters": [
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "title": "Response Broken Refs V1 Admin Kb Rule Refs Broken Get"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Broken Refs"
   }
  },
...
  "/v1/kb/clauses/{code}/{clause_no}": {
   "get": {
    "operationId": "get_clause_v1_kb_clauses__code___clause_no__get",
    "parameters": [
     {
      "in": "path",
      "name": "code",
      "required": true,
      "schema": {
       "title": "Code",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "clause_no",
      "required": true,
      "schema": {
       "title": "Clause No",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "authorization",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "title": "Response Get Clause V1 Kb Clauses  Code   Clause No  Get"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Clause"
   }
  },
...
 }
}